% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridstackeR.R
\name{grid_stack_item}
\alias{grid_stack_item}
\title{Grid Stack Item}
\usage{
grid_stack_item(..., height = 4, width = 4, x = 0, y = 0)
}
\arguments{
\item{...}{content to include in the grid stack item}

\item{height}{the height in rows}

\item{width}{the width in columns usually 1-12}

\item{x}{the x-position}

\item{y}{the y-position}
}
\description{
This is a wrapper for the individual items to be displayed in the \link{grid_stack}
}
